[#267] Move tests to more submodules
Problem: Right now we have some test modules that have scenarios as well
as test trees. But we would like to keep them separate, and split big
modules into smaller ones.

Solution: Move scenarios to separate mode, and split big modules into
smaller ones.
