+++ haskell/baseDAO-ligo-meta.cabal
@@ -87,10 +87,14 @@ test-suite baseDAO-test
       Test.Ligo.BaseDAO.Integrational.Common
       Test.Ligo.BaseDAO.Integrational.Proposal
       Test.Ligo.BaseDAO.Management
+      Test.Ligo.BaseDAO.Management.TransferOwnership
       Test.Ligo.BaseDAO.OffChainViews
       Test.Ligo.BaseDAO.Proposal
       Test.Ligo.BaseDAO.Proposal.Config
-      Test.Ligo.BaseDAO.Proposal.Proposal
+      Test.Ligo.BaseDAO.Proposal.Flush
+      Test.Ligo.BaseDAO.Proposal.Propose
+      Test.Ligo.BaseDAO.Proposal.Quorum
+      Test.Ligo.BaseDAO.Proposal.Tokens
       Test.Ligo.BaseDAO.Proposal.Vote
       Test.Ligo.BaseDAO.Token
       Test.Ligo.BaseDAO.Token.BalanceOf
